Create Find Cracker.
Someone was hacking the score. Each student's record is given as an array The objects in the array are given again as an array of scores for each name and total score. ex. The scores for each grade are:
A: 30 points
B: 20 points
C: 10 points
D: 5 points
Everything else: 0 points
If there are 5 or more courses and all courses have a grade of B or above, additional 20 points are awarded. After all the calculations, the total score should be capped at 200 points.
Returns the name of the hacked name as an array when scoring with this rule.